requirements.txt
crewai==1.7.2
crewai-tools==1.7.2
intersystems-irispython==5.3.1
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-dotenv==1.1.1
pydantic==2.11.10
